+++ fixtures/partytracks/src/server/index.ts
@@ -1,8 +1,20 @@
-// biome-ignore lint/complexity/noBannedTypes: <explanation>
-type Env = {};
+import { Hono } from "hono";
+import { proxyToCallsApi } from "partytracks/server";
 
-export default {
-  async fetch(request: Request, env: Env, ctx: ExecutionContext) {
-    return new Response("Hello World");
-  }
-} satisfies ExportedHandler<Env>;
+type Bindings = {
+  CALLS_APP_ID: string;
+  CALLS_APP_TOKEN: string;
+};
+
+const app = new Hono<{ Bindings: Bindings }>();
+
+app.all("/api/calls/*", (c) =>
+  proxyToCallsApi({
+    replaceProxyPathname: "/api/calls",
+    appId: c.env.CALLS_APP_ID,
+    token: c.env.CALLS_APP_TOKEN,
+    request: c.req.raw
+  })
+);
+
+export default app;