+++ packages/partytracks/src/client/index.ts
@@ -1,3 +1,4 @@
 export { PartyTracks } from "./PartyTracks";
+export { setLogLevel } from "./logging";
 export type { PartyTracksConfig, ApiHistoryEntry } from "./PartyTracks";
 export type { TrackMetadata } from "./callsTypes";