+++ packages/partytracks/src/debug.ts
@@ -1,8 +1,0 @@
-import { tap } from "rxjs";
-
-export function debugTap<T>(message: string) {
-  return tap<T>({
-    next: (...args) => console.log(message, ...args),
-    complete: () => console.log("COMPLETED ", message)
-  });
-}